updated the player model
changed the image from a rectangle to a sprite

+++ proj/src/devices/video_card/video.h
@@ -38,6 +38,6 @@ int (video_draw_row)(uint16_t x, uint16_t y, uint16_t len, uint32_t color);
 int (video_draw_col)(uint16_t x, uint16_t y, uint16_t len, uint32_t color);
 int (video_draw_rectangle)(uint16_t x, uint16_t y, uint16_t width, uint16_t height, uint32_t color);
 
-int (video_draw_sprite)(uint16_t x, uint16_t y, uint16_t width, uint16_t height, uint32_t* colors);
+int (video_draw_sprite)(uint16_t x, uint16_t y, uint16_t width, uint16_t height, uint16_t* colors);
 
 #endif // _LCOM_VIDEO_H_
